Register.java
In this form whenever a new candidate comes to apply for a job, he/she has to register into the campus recruitment system. he/she will then be registered for application of the post for which he/she applied. On the basis of the entries made in the registration form a mail will be sent to the their respective address about the vacancy of the post and the schedule of their interview.
        when the entries are made by the candidate , information is stored in the database maintain by the admin.